# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Global Tongue API Documentation",
        "description": "Heirarchy of API | Courses -> Levels -> Units -> Lessons -> Exercises",
        "contact": {
            "email": "gemeelijah@gmail.com"
        },
        "license": {
            "name": "www.globaltongueedu.com",
            "url": "https://globaltongueedu.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/public/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Show Course Details",
                "description": "Show Course Details",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Course Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update Course",
                "description": "Update Course",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "English for Beginner"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Here are the biggest enterprise technology acquisitions of 2021 so far, in reverse chronological order."
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Beginner"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "free"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Course"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete Course",
                "description": "Delete Course",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Course"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/v1/public/api/levels": {
            "get": {
                "tags": [
                    "Levels"
                ],
                "summary": "Get Level List",
                "description": "Get Level List as Array",
                "operationId": "indexLevel",
                "responses": {
                    "200": {
                        "description": "Get Level List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Levels"
                ],
                "summary": "Create New Level",
                "description": "Create New Level",
                "operationId": "storeLevel",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Spanish Level 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Learn words and phrases for greetings and introductions, eating at a restaurant, shopping, family, and travel. Study professions, hobbies, pronunciation of -r versus -rr, and subject pronouns and learn when to use tú versus usted."
                                    },
                                    "tag": {
                                        "type": "string",
                                        "example": "Ser, Gender, Gustar, Estar, Plurals,"
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Level 1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "free"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create New Level"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/levels/view/all": {
            "get": {
                "tags": [
                    "Levels"
                ],
                "summary": "All Levels - Publicly Accessible",
                "description": "All Levels - Publicly Accessible",
                "operationId": "indexAllLevel",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Levels - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/v1/public/api/levels/by/{course_id}": {
            "get": {
                "tags": [
                    "Levels"
                ],
                "summary": "Get Levels by Course ID",
                "description": "Get list of levels associated with a specific course",
                "operationId": "getLevelsByCourseID",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Levels for the specified Course ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No Levels found for the specified Course ID"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/levels/{id}": {
            "get": {
                "tags": [
                    "Levels"
                ],
                "summary": "Show Level Details",
                "description": "Show Level Details",
                "operationId": "showLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Level Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Levels"
                ],
                "summary": "Update Level",
                "description": "Update Level",
                "operationId": "updateLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Spanish Level 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Learn words and phrases for greetings and introductions, eating at a restaurant, shopping, family, and travel. Study professions, hobbies, pronunciation of -r versus -rr, and subject pronouns and learn when to use tú versus usted."
                                    },
                                    "tag": {
                                        "type": "string",
                                        "example": "Ser, Gender, Gustar, Estar, Plurals,"
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Level 1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "free"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Level"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Levels"
                ],
                "summary": "Delete Level",
                "description": "Delete Level",
                "operationId": "destroyLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Level"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}